Empire achievement wrapper
Inching closer to something usable.

+++ src/main/java/de/geolykt/starloader/apimixins/StateActorMixins.java
@@ -57,7 +57,7 @@ public abstract class StateActorMixins extends ActorMixins implements de.geolykt
         // SLAPI: use distSq instead of dist - avoiding a Math#sqrt call
         if (this.location == null || getDistSq(this.location) > 0.01D) {
             var oldLocation = this.location;
-            this.location = Space.findStarNear(this.getX(), this.getY(), (double) (Space.getMaxX() * 200.0F), (snoddasmannen.galimulator.Empire) null);
+            this.location = Space.findStarNear(this.getX(), this.getY(), Space.getMaxX() * 200.0F, (snoddasmannen.galimulator.Empire) null);
             if (oldLocation != this.location) {
                 this.guide.arrivedAt(this.location);
             }
